+++ api/src/modules/authorization/permissions/permissions.entity.ts
@@ -1,19 +1,21 @@
-import {
-  BaseEntity,
-  Entity,
-  JoinTable,
-  ManyToMany,
-  PrimaryColumn,
-} from 'typeorm';
+import { BaseEntity, Entity, ManyToMany, PrimaryColumn } from 'typeorm';
 import { PERMISSIONS } from 'modules/authorization/permissions/permissions.enum';
 import { Role } from 'modules/authorization/roles/role.entity';
+import { ApiProperty } from '@nestjs/swagger';
 
 @Entity('permissions')
 export class Permission extends BaseEntity {
-  @PrimaryColumn({ type: 'varchar', enum: PERMISSIONS })
+  @ApiProperty()
+  @PrimaryColumn({
+    type: 'varchar',
+    enum: PERMISSIONS,
+    unique: true,
+    default: PERMISSIONS.CAN_CREATE_SCENARIO,
+  })
   action!: PERMISSIONS;
 
   @ManyToMany(() => Role, (role: Role) => role.permissions)
-  @JoinTable()
   roles: Role[];
+
+  //TODO: Add a default flag? It might be useful
 }