+++ client/src/components/forms/input/component.tsx
@@ -13,7 +13,11 @@ const Input = forwardRef<HTMLInputElement, InputProps>(
         {icon && (
           <div className="absolute inset-y-0 left-0 flex items-center pl-3 pointer-events-none">
             {cloneElement(icon, {
-              className: classnames('w-4 h-4 text-gray-500', icon.props?.className),
+              className: classnames(
+                'w-4 h-4',
+                inputProps.disabled ? 'text-gray-300' : 'text-gray-500',
+                icon.props?.className,
+              ),
             })}
           </div>
         )}
@@ -21,7 +25,7 @@ const Input = forwardRef<HTMLInputElement, InputProps>(
           className={classnames(
             'block w-full px-4 py-2.5 leading-5 rounded-md bg-white text-sm border placeholder:text-gray-200 focus:outline-none focus:ring-0',
             error ? 'border-red-400 focus:border-red-400' : 'border-gray-200 focus:border-navy-400',
-            inputProps.disabled ? 'text-gray-200' : 'text-gray-900',
+            inputProps.disabled ? 'bg-gray-300/20 text-gray-300' : 'text-gray-900',
             {
               'pl-8': icon,
               'pr-10': unit && !error,